requirements.txt
streamlit
langchain
langchain-openai
langgraph
openai
chromadb
folium
streamlit-folium
pandas
geopy
python-dotenv
plotly
requests
numpy
scipy
